Rename boot auth file to be more intentful
As [discussed with CodeRabbit][1].

[1]: https://github.com/ProteusHK/the-digger/issues/849#issuecomment-3921357882
