+++ src/main.ts
@@ -1,5 +1,5 @@
 import './assets/main.css';
-import './boot/auth-header'; // <-- attach Authorization on all /api requests (and persist login token)
+import './boot/auth-bearer-token'; // <-- attach Authorization on all /api requests (and persist login token)
 
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
